requirements.txt
aiohttp>=3.8.0
pyrogram==2.0.106
googletrans==4.0.0-rc1
python-dotenv==1.0.0
moviepy==1.0.3
rich==13.7.0
colorama==0.4.6
requests==2.31.0
tgcrypto
psutil>=5.9.0